move injectProjectDir to legacy.go
